Separate update python packages command
